sqlite3_prepare_v2: minor fixup
The sqlite3_prepare_v2 takes a length parameter that specifies the
maximum number of characters that the parser will look at while
preparing the statement. If it is set to -1, the parser just terminates
when it reaches a null string character. This means that by passing
strlen(cstr), we are effectively calling strlen twice.

With this change, the insert performance test, which calls prepare
100000 times is roughly 3-4% faster on average.

Reviewed-by: sean.c.rhea@gmail.com