change query interface, add prepared statements
This commit makes three related changes.

1. The old interface for the query function relied on the user to
iterate completely through the result, or else the underlying statement
on the C heap would not be deallocated the garbage collector called the
result iterator's finalize function.

In the new interface, instead of returning an iterator, query takes
function from Iterator => R, so that query itself deallocates the
statement object when the iterator is no longer in scope.

Before, the following code might have left a statement allocated for an
arbitrary period:

  val row = db.query(sql).next

The new version will deallocate the statement immediately. It looks
like this:

  val row = db.query(sql)(_.next)

2. In client code using the old version, one often wrote code like this:

  for (row <- db.query(sql)) {
    foobar(row)
  }

To ease porting to this new version without reindenting much code,
we provide the foreachRow convenience function:

  db.foreachRow(sql) {
    foobar(row)
  }

3. The new prepare function can be used to bind different values to the
same statement, resulting in significant performance improvements when
executing many similar queries. (In practice, prepared statements are
also less susceptible to sql injections.) It is used like this:

  val sql = "select * from users where id = ?"
  db.prepare(sql) { stmt =>
    val user1 = stmt.query(SqlInt(1))(_.next)
    val user2 = stmt.query(SqlInt(2))(_.next)
    // ...
  }

Reviewed-by: sean.c.rhea@gmail.com