add test that prepared statements correctly reset
It is currently possible to get a prepared statement into a weird
state because it is not always reset. This commit adds a failing
test for two separate queries against a single prepared statement.
The prepared statement interface will be changed in the next commit
to prevent this bug. There will also be a revised version of this
test for the new interface.