make prepared statements safer
We change the interface to prepared statement queries so that
they are wrapped in a call to reset. This ensures that we won't
end up with a prepared statement that is in an unexpected
state because reset wasn't called.

This change also requires a slight modification to the one-off
query code, though not to the interface itself.

The test is updated for the new prepared statement interface
query and now passes.