Separate TEXT and BLOB types
In most cases it is perfectly fine to interchange use BLOB columns to
represent TEXT columns. Occasionally, however, you will get strange
behavior when comparing BLOB and TEXT columns. The reason that there was
a single type of string column was for compatibility with a foreign
table that was added as a virtual table to Sqlite that did not
distinguish between string types. It is a real headache to debug the
subtle errors that can crop up when you need to compare the different
types of string columns, so we are making them distinct in Scalaqlite
once again.

With this change, queries that return columns of TEXT type will now
return a SqlText object, not a SqlBlob as before. This usually will be a
functional no op because the toString method of SqlBlob should return
the same string as the toString method of the SqlText object that is
instantiated with a ByteArray.