SqliteDb.scala: add mapRows and getRows methods
One often wants to run a query and cache the results in memory. These
two methods should make this a bit more straightforward. The mapRows
and getRows methods are just syntactic sugar to reduce the indentation
by one level whenever one wants a set of all the rows returned by a
query or if one wants to apply a mapping to these rows and cache the
results.

Bonus: I removed the ListBuffer import because it isn't used anywhere.