SqliteDbSpec.scala: replace type affinity test
There was an existing test that checked that BLOB columns were bound as
TEXT that was an artifact of when we only had a single string type for
both BLOBs and TEXTs. That test started failing with the change to bind
BLOBs as BLOBs and to return TEXT columns as TEXT. I've replaced that
test with a new test that verifies that a number of BLOB/BLOB,
BLOB/TEXT and TEXT/TEXT comparisons work as expected. It tests
comparisons between sets of bound parameters, comparisons between bound
parameters and constants and comparisons between db columns.